requirements.txt
pygame-ce
pandas
